+++ plugins/crepuscularity-components/packages/flutter/test/themes_test.dart
@@ -9,6 +9,8 @@ void main() {
       'night',
       'chalk',
       'aurora',
+      'dawn',
+      'zinc',
     ]));
   });